+++ exporters/otlp/otlpmetric/exporter_test.go
@@ -31,6 +31,7 @@ import (
 	"go.opentelemetry.io/otel/attribute"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/number"
+	"go.opentelemetry.io/otel/metric/sdkapi"
 	metricsdk "go.opentelemetry.io/otel/sdk/export/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
 	commonpb "go.opentelemetry.io/proto/otlp/common/v1"
@@ -98,7 +99,7 @@ func (m *checkpointSet) ForEach(_ metricsdk.ExportKindSelector, fn func(metricsd
 
 type record struct {
 	name     string
-	iKind    metric.InstrumentKind
+	iKind    sdkapi.InstrumentKind
 	nKind    number.Kind
 	resource *resource.Resource
 	opts     []metric.InstrumentOption
@@ -184,7 +185,7 @@ func TestNoGroupingExport(t *testing.T) {
 		[]record{
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				nil,
 				nil,
@@ -192,7 +193,7 @@ func TestNoGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				nil,
 				nil,
@@ -239,7 +240,7 @@ func TestNoGroupingExport(t *testing.T) {
 func TestValuerecorderMetricGroupingExport(t *testing.T) {
 	r := record{
 		"valuerecorder",
-		metric.ValueRecorderInstrumentKind,
+		sdkapi.ValueRecorderInstrumentKind,
 		number.Int64Kind,
 		nil,
 		nil,
@@ -290,7 +291,7 @@ func TestValuerecorderMetricGroupingExport(t *testing.T) {
 func TestCountInt64MetricGroupingExport(t *testing.T) {
 	r := record{
 		"int64-count",
-		metric.CounterInstrumentKind,
+		sdkapi.CounterInstrumentKind,
 		number.Int64Kind,
 		nil,
 		nil,
@@ -340,7 +341,7 @@ func TestCountInt64MetricGroupingExport(t *testing.T) {
 func TestCountFloat64MetricGroupingExport(t *testing.T) {
 	r := record{
 		"float64-count",
-		metric.CounterInstrumentKind,
+		sdkapi.CounterInstrumentKind,
 		number.Float64Kind,
 		nil,
 		nil,
@@ -394,7 +395,7 @@ func TestResourceMetricGroupingExport(t *testing.T) {
 		[]record{
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				nil,
@@ -402,7 +403,7 @@ func TestResourceMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				nil,
@@ -410,7 +411,7 @@ func TestResourceMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				nil,
@@ -418,7 +419,7 @@ func TestResourceMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstB,
 				nil,
@@ -512,7 +513,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 		[]record{
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				countingLib1,
@@ -520,7 +521,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				countingLib2,
@@ -528,7 +529,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				countingLib1,
@@ -536,7 +537,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				countingLib1,
@@ -544,7 +545,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstA,
 				summingLib,
@@ -552,7 +553,7 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 			},
 			{
 				"int64-count",
-				metric.CounterInstrumentKind,
+				sdkapi.CounterInstrumentKind,
 				number.Int64Kind,
 				testInstB,
 				countingLib1,
@@ -688,16 +689,16 @@ func TestResourceInstLibMetricGroupingExport(t *testing.T) {
 func TestStatelessExportKind(t *testing.T) {
 	type testcase struct {
 		name           string
-		instrumentKind metric.InstrumentKind
+		instrumentKind sdkapi.InstrumentKind
 		aggTemporality metricpb.AggregationTemporality
 		monotonic      bool
 	}
 
 	for _, k := range []testcase{
-		{"counter", metric.CounterInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_DELTA, true},
-		{"updowncounter", metric.UpDownCounterInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_DELTA, false},
-		{"sumobserver", metric.SumObserverInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_CUMULATIVE, true},
-		{"updownsumobserver", metric.UpDownSumObserverInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_CUMULATIVE, false},
+		{"counter", sdkapi.CounterInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_DELTA, true},
+		{"updowncounter", sdkapi.UpDownCounterInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_DELTA, false},
+		{"sumobserver", sdkapi.SumObserverInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_CUMULATIVE, true},
+		{"updownsumobserver", sdkapi.UpDownSumObserverInstrumentKind, metricpb.AggregationTemporality_AGGREGATION_TEMPORALITY_CUMULATIVE, false},
 	} {
 		t.Run(k.name, func(t *testing.T) {
 			runMetricExportTests(